requirements.txt
streamlit==1.55.0
pandas==2.3.3
google-generativeai==0.8.5
openai==2.29.0
python-dotenv==1.2.2
